+++ CHANGELOG.md
@@ -2,6 +2,34 @@
 
 All notable changes to this project will be documented in this file.
 
+### [1.1.3](https://github.com/wandb/terraform-kubernetes-wandb/compare/v1.1.2...v1.1.3) (2022-06-16)
+
+
+### Bug Fixes
+
+* Change connection string to use kubernetes secret ([#13](https://github.com/wandb/terraform-kubernetes-wandb/issues/13)) ([5093996](https://github.com/wandb/terraform-kubernetes-wandb/commit/5093996aa448942b9dfe14f2e3d98ef7451f62ea))
+
+### [1.1.2](https://github.com/wandb/terraform-kubernetes-wandb/compare/v1.1.1...v1.1.2) (2022-05-10)
+
+
+### Bug Fixes
+
+* Remove count on config map ([#12](https://github.com/wandb/terraform-kubernetes-wandb/issues/12)) ([ed00777](https://github.com/wandb/terraform-kubernetes-wandb/commit/ed00777d27b0c6bfa3f8b5cd1cb23012ecfffe0d))
+
+### [1.1.1](https://github.com/wandb/terraform-kubernetes-wandb/compare/v1.1.0...v1.1.1) (2022-05-05)
+
+
+### Bug Fixes
+
+* Update cert path ([#11](https://github.com/wandb/terraform-kubernetes-wandb/issues/11)) ([fcf8aef](https://github.com/wandb/terraform-kubernetes-wandb/commit/fcf8aef30f82527ced96470c2fd98cc6205eaeeb))
+
+## [1.1.0](https://github.com/wandb/terraform-kubernetes-wandb/compare/v1.0.3...v1.1.0) (2022-05-04)
+
+
+### Features
+
+* Add option to mount redis cert ([#10](https://github.com/wandb/terraform-kubernetes-wandb/issues/10)) ([4f51ac4](https://github.com/wandb/terraform-kubernetes-wandb/commit/4f51ac43635a3094a7fe65c25a66336771e7a7f1))
+
 ### [1.0.3](https://github.com/wandb/terraform-kubernetes-wandb/compare/v1.0.2...v1.0.3) (2022-04-20)